Add importable browser module
Moves the module registration logic to a new pkg called browser.
